修改授权功能

+++ actix-test/static/test_build_js.sh
@@ -1,13 +1,17 @@
+#!/bin/bash
+
 unset http_proxy
 unset https_proxy
 
-cd /home/jeremy/work/ClionProjects/self/demo/self-wlib
-rm -rf self-wlib-1.0.0.tgz 
-webpack index.js
-npm pack
-cp self-wlib-1.0.0.tgz /home/jeremy/work/ClionProjects/parts-learning/actix-test/static
+if [ $1 = "true" ];then
+	cd /home/jeremy/work/ClionProjects/self/demo/self-wlib
+	rm -rf self-wlib-1.0.0.tgz
+	webpack index.js
+	npm pack
+	cp self-wlib-1.0.0.tgz /home/jeremy/work/ClionProjects/parts-learning/actix-test/static
+	cd /home/jeremy/work/ClionProjects/parts-learning/actix-test/static
+	npm install self-wlib-1.0.0.tgz
+fi
 
-cd /home/jeremy/work/ClionProjects/parts-learning/actix-test/static
-npm install self-wlib-1.0.0.tgz
 webpack index.js
 echo "update js file is over!"
